pyproject.toml
[project]
name = "newsnewt"
version = "0.1.0"
description = "Archive and extract news articles from URLs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "archivenow>=2020.7.18.12.19.44",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "trafilatura>=1.7.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "setuptools>=75.0.0",
]
[project.optional-dependencies]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.6",
    "pyright>=1.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newsnewt"]
